feat: Implement Admin Prompt Store with modals for viewing, creating, and editing prompts
- Added AdminPromptExpandModal for displaying prompt details
- Created AdminPromptStore component to manage prompts with filtering and searching capabilities
- Integrated AdminPromptCreateModal and AdminPromptEditModal for creating and editing prompts
- Implemented styles for the prompt store and modals
- Added functionality for copying prompts to clipboard and deleting prompts with confirmation
- Enhanced user experience with loading states and error handling
