+++ src/app/admin/components/workflows/AdminPromptStore/AdminPromptEditModal.module.scss
@@ -1,3 +1,6 @@
+@use "sass:color";
+@use "@/app/_common/_variables" as *;
+
 .modalOverlay {
     position: fixed;
     top: 0;
@@ -13,11 +16,11 @@
 }
 
 .modalContent {
-    background: var(--background-primary);
+    background: $white;
     border-radius: 12px;
-    width: 90%;
-    max-width: 700px;
-    max-height: 90vh;
+    width: 95%;
+    max-width: 1000px;
+    max-height: 85vh;
     display: flex;
     flex-direction: column;
     box-shadow: 0 20px 25px -5px rgba(0, 0, 0, 0.1), 0 10px 10px -5px rgba(0, 0, 0, 0.04);
@@ -27,13 +30,14 @@
     display: flex;
     align-items: center;
     justify-content: space-between;
-    padding: 24px;
-    border-bottom: 1px solid var(--border-color);
+    padding: 20px 28px;
+    border-bottom: 1px solid $gray-200;
+    flex-shrink: 0;
 }
 
 .modalTitle {
     margin: 0;
-    color: var(--text-primary);
+    color: $gray-900;
     font-size: 20px;
     font-weight: 600;
 }
@@ -46,15 +50,15 @@
     height: 36px;
     border: none;
     border-radius: 6px;
-    background: var(--background-secondary);
-    color: var(--text-secondary);
+    background: $gray-100;
+    color: $gray-600;
     font-size: 24px;
     cursor: pointer;
     transition: all 0.2s ease;
 
     &:hover:not(:disabled) {
-        background: var(--background-tertiary);
-        color: var(--text-primary);
+        background: $gray-200;
+        color: $gray-900;
     }
 
     &:disabled {
@@ -66,77 +70,85 @@
 .modalBody {
     flex: 1;
     overflow-y: auto;
-    padding: 24px;
+    padding: 24px 28px;
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    gap: 24px;
+
+    @media (max-width: 768px) {
+        grid-template-columns: 1fr;
+    }
 }
 
 .formGroup {
-    margin-bottom: 20px;
+    margin-bottom: 0;
 
-    &:last-child {
-        margin-bottom: 0;
+    &.fullWidth {
+        grid-column: 1 / -1;
     }
 }
 
 .label {
     display: block;
-    margin-bottom: 8px;
-    color: var(--text-primary);
-    font-size: 14px;
+    margin-bottom: 6px;
+    color: $gray-900;
+    font-size: 13px;
     font-weight: 600;
 }
 
 .required {
-    color: #ef4444;
+    color: $primary-red;
 }
 
 .input,
 .textarea {
     width: 100%;
-    padding: 12px;
-    border: 1px solid var(--border-color);
+    padding: 10px 12px;
+    border: 1px solid $gray-300;
     border-radius: 6px;
-    background: var(--background-secondary);
-    color: var(--text-primary);
-    font-size: 14px;
+    background: $white;
+    color: $gray-900;
+    font-size: 13px;
     font-family: inherit;
     transition: all 0.2s ease;
 
     &:focus {
         outline: none;
-        border-color: var(--primary-color);
-        box-shadow: 0 0 0 3px rgba(59, 130, 246, 0.1);
+        border-color: $primary-blue;
+        box-shadow: 0 0 0 3px rgba($primary-blue, 0.1);
     }
 
     &.error {
-        border-color: #ef4444;
+        border-color: $primary-red;
     }
 
     &:disabled {
         opacity: 0.6;
         cursor: not-allowed;
+        background: $gray-50;
     }
 
     &::placeholder {
-        color: var(--text-tertiary);
+        color: $gray-400;
     }
 }
 
 .textarea {
     resize: vertical;
-    min-height: 150px;
+    min-height: 200px;
     line-height: 1.6;
 }
 
 .errorMessage {
     margin-top: 6px;
-    color: #ef4444;
+    color: $primary-red;
     font-size: 12px;
 }
 
 .charCount {
-    margin-top: 6px;
-    color: var(--text-tertiary);
-    font-size: 12px;
+    margin-top: 4px;
+    color: $gray-500;
+    font-size: 11px;
     text-align: right;
 }
 
@@ -150,26 +162,26 @@
     display: flex;
     align-items: center;
     justify-content: center;
-    gap: 8px;
-    padding: 12px;
-    border: 1px solid var(--border-color);
+    gap: 6px;
+    padding: 10px;
+    border: 1px solid $gray-300;
     border-radius: 6px;
-    background: var(--background-secondary);
-    color: var(--text-secondary);
-    font-size: 14px;
+    background: $white;
+    color: $gray-700;
+    font-size: 13px;
     font-weight: 500;
     cursor: pointer;
     transition: all 0.2s ease;
 
     &:hover:not(:disabled) {
-        background: var(--background-tertiary);
-        border-color: var(--primary-color);
+        background: $gray-50;
+        border-color: $gray-400;
     }
 
     &.active {
-        background: var(--primary-color);
-        border-color: var(--primary-color);
-        color: white;
+        background: $primary-blue;
+        border-color: $primary-blue;
+        color: $white;
     }
 
     &:disabled {
@@ -183,8 +195,8 @@
     align-items: center;
     gap: 10px;
     cursor: pointer;
-    color: var(--text-primary);
-    font-size: 14px;
+    color: $gray-900;
+    font-size: 13px;
     font-weight: 500;
 }
 
@@ -198,49 +210,48 @@
     justify-content: center;
     width: 20px;
     height: 20px;
-    border: 2px solid var(--border-color);
+    border: 2px solid $gray-300;
     border-radius: 4px;
-    background: var(--background-secondary);
+    background: $white;
     transition: all 0.2s ease;
 
     .checkIcon {
-        color: white;
+        color: $white;
         font-size: 16px;
     }
 }
 
 .checkbox:checked + .checkboxText {
-    background: var(--primary-color);
-    border-color: var(--primary-color);
+    background: $primary-blue;
+    border-color: $primary-blue;
 }
 
 .helpText {
-    margin-top: 6px;
+    margin-top: 4px;
     margin-left: 30px;
-    color: var(--text-tertiary);
-    font-size: 12px;
-    line-height: 1.5;
+    color: $gray-500;
+    font-size: 11px;
+    line-height: 1.4;
 }
 
 .infoBox {
-    padding: 16px;
-    border: 1px solid var(--border-color);
+    padding: 14px 16px;
+    background: $gray-50;
+    border: 1px solid $gray-200;
     border-radius: 8px;
-    background: var(--background-secondary);
-    margin-top: 20px;
 }
 
 .infoTitle {
-    margin-bottom: 12px;
-    color: var(--text-primary);
-    font-size: 14px;
+    margin-bottom: 10px;
+    color: $gray-900;
+    font-size: 13px;
     font-weight: 600;
 }
 
 .infoContent {
     display: flex;
     flex-direction: column;
-    gap: 8px;
+    gap: 6px;
 }
 
 .infoItem {
@@ -250,14 +261,16 @@
 }
 
 .infoLabel {
-    color: var(--text-tertiary);
-    font-size: 13px;
+    min-width: 90px;
+    color: $gray-600;
+    font-size: 12px;
     font-weight: 500;
 }
 
 .infoValue {
-    color: var(--text-secondary);
-    font-size: 13px;
+    color: $gray-900;
+    font-size: 12px;
+    font-weight: 400;
 }
 
 .modalFooter {
@@ -265,24 +278,26 @@
     align-items: center;
     justify-content: flex-end;
     gap: 12px;
-    padding: 20px 24px;
-    border-top: 1px solid var(--border-color);
+    padding: 16px 28px;
+    border-top: 1px solid $gray-200;
+    flex-shrink: 0;
 }
 
 .cancelButton {
-    padding: 10px 24px;
-    border: 1px solid var(--border-color);
+    min-width: 100px;
+    padding: 11px 24px;
+    border: 1px solid $gray-300;
     border-radius: 6px;
-    background: var(--background-secondary);
-    color: var(--text-secondary);
+    background: $white;
+    color: $gray-700;
     font-size: 14px;
     font-weight: 500;
     cursor: pointer;
     transition: all 0.2s ease;
 
     &:hover:not(:disabled) {
-        background: var(--background-tertiary);
-        border-color: var(--primary-color);
+        background: $gray-50;
+        border-color: $gray-400;
     }
 
     &:disabled {
@@ -292,18 +307,19 @@
 }
 
 .submitButton {
-    padding: 10px 24px;
+    min-width: 100px;
+    padding: 11px 24px;
     border: none;
     border-radius: 6px;
-    background: var(--primary-color);
-    color: white;
+    background: $primary-blue;
+    color: $white;
     font-size: 14px;
     font-weight: 500;
     cursor: pointer;
     transition: all 0.2s ease;
 
     &:hover:not(:disabled) {
-        background: #2563eb;
+        background: color.adjust($primary-blue, $lightness: -5%);
     }
 
     &:disabled {
