+++ src/app/admin/components/workflows/AdminPromptStore/AdminPromptExpandModal.module.scss
@@ -1,3 +1,6 @@
+@use "sass:color";
+@use "@/app/_common/_variables" as *;
+
 .modalOverlay {
     position: fixed;
     top: 0;
@@ -13,7 +16,7 @@
 }
 
 .modalContent {
-    background: var(--background-primary);
+    background: $white;
     border-radius: 12px;
     width: 90%;
     max-width: 800px;
@@ -28,12 +31,12 @@
     align-items: center;
     justify-content: space-between;
     padding: 24px;
-    border-bottom: 1px solid var(--border-color);
+    border-bottom: 1px solid $gray-200;
 }
 
 .modalTitle {
     margin: 0;
-    color: var(--text-primary);
+    color: $gray-900;
     font-size: 20px;
     font-weight: 600;
     flex: 1;
@@ -48,15 +51,15 @@
     height: 36px;
     border: none;
     border-radius: 6px;
-    background: var(--background-secondary);
-    color: var(--text-secondary);
+    background: $gray-100;
+    color: $gray-600;
     font-size: 24px;
     cursor: pointer;
     transition: all 0.2s ease;
 
     &:hover {
-        background: var(--background-tertiary);
-        color: var(--text-primary);
+        background: $gray-200;
+        color: $gray-900;
     }
 }
 
@@ -82,8 +85,8 @@
     letter-spacing: 0.5px;
 
     &.language {
-        background: rgba(59, 130, 246, 0.1);
-        color: #3b82f6;
+        background: rgba($primary-blue, 0.1);
+        color: $primary-blue;
     }
 
     &.template {
@@ -115,7 +118,7 @@
 
 .sectionTitle {
     margin: 0;
-    color: var(--text-primary);
+    color: $gray-900;
     font-size: 16px;
     font-weight: 600;
 }
@@ -125,19 +128,19 @@
     align-items: center;
     gap: 6px;
     padding: 8px 16px;
-    border: 1px solid var(--border-color);
+    border: 1px solid $gray-300;
     border-radius: 6px;
-    background: var(--background-secondary);
-    color: var(--text-secondary);
+    background: $white;
+    color: $gray-700;
     font-size: 14px;
     font-weight: 500;
     cursor: pointer;
     transition: all 0.2s ease;
 
     &:hover {
-        background: var(--primary-color);
-        border-color: var(--primary-color);
-        color: white;
+        background: $primary-blue;
+        border-color: $primary-blue;
+        color: $white;
     }
 
     .copyIcon {
@@ -147,10 +150,10 @@
 
 .contentBox {
     padding: 16px;
-    border: 1px solid var(--border-color);
+    border: 1px solid $gray-200;
     border-radius: 8px;
-    background: var(--background-secondary);
-    color: var(--text-primary);
+    background: $gray-50;
+    color: $gray-900;
     font-size: 14px;
     line-height: 1.8;
     white-space: pre-wrap;
@@ -183,7 +186,7 @@
     display: flex;
     align-items: center;
     gap: 6px;
-    color: var(--text-tertiary);
+    color: $gray-500;
     font-size: 12px;
     font-weight: 600;
     text-transform: uppercase;
@@ -195,7 +198,7 @@
 }
 
 .metaValue {
-    color: var(--text-primary);
+    color: $gray-900;
     font-size: 14px;
     font-weight: 500;
 }
@@ -205,23 +208,23 @@
     align-items: center;
     justify-content: flex-end;
     padding: 20px 24px;
-    border-top: 1px solid var(--border-color);
+    border-top: 1px solid $gray-200;
 }
 
 .closeFooterButton {
     padding: 10px 24px;
-    border: 1px solid var(--border-color);
+    border: 1px solid $gray-300;
     border-radius: 6px;
-    background: var(--background-secondary);
-    color: var(--text-secondary);
+    background: $white;
+    color: $gray-700;
     font-size: 14px;
     font-weight: 500;
     cursor: pointer;
     transition: all 0.2s ease;
 
     &:hover {
-        background: var(--background-tertiary);
-        border-color: var(--primary-color);
-        color: var(--text-primary);
+        background: $gray-50;
+        border-color: $gray-400;
+        color: $gray-900;
     }
 }
